+++ M15_JocDeDaus/src/main/java/com/ITAcademy/M15_JocDeDaus/Services/IGameService.java
@@ -1,6 +1,5 @@
 package com.ITAcademy.M15_JocDeDaus.Services;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 import com.ITAcademy.M15_JocDeDaus.DTO.GameDTO;
@@ -10,7 +9,7 @@ public interface IGameService {
 	public GameDTO playGame(long player_id);
 	public List<GameDTO> gamesByPlayer(long player_id);
 	public void deleteGamesByPlayer(long player_id);
-	public BigDecimal calculateWinRate(long player_id);
+	public double calculateWinRate(long player_id);
 	public Boolean checkNoGames();
